fix(protocol/robot): 优化状态上报逻辑并修复电量类型
1. 将电量字段类型从int改为float64适配实际上报场景
2. 新增状态快照与上报过滤逻辑，仅在状态变化超过阈值时才发送消息
3. 调整错误处理为非致命日志，避免程序异常退出
4. 增加初始状态上报，确保连接建立后立即发送当前状态

+++ protocol/message.go
@@ -26,7 +26,6 @@ func newHeader() Header {
 
 // ============ 状态常量（协议层用字符串，跨语言友好）============
 
-
 const (
 	StateIdle     = "IDLE"
 	StateMoving   = "MOVING"
@@ -53,7 +52,7 @@ type StatusBody struct {
 	ID      string  `json:"id"`                // 设备id
 	X       float64 `json:"x"`                 // x 坐标
 	Y       float64 `json:"y"`                 // y 坐标
-	Battery int     `json:"battery"`           // 电量
+	Battery float64 `json:"battery"`           // 电量
 	State   string  `json:"state"`             // 状态
 	Speed   float64 `json:"speed"`             // 速度
 	TaskID  string  `json:"task_id,omitempty"` // 当前执行的任务ID，空则省略
